Relocated tests for credentials files
